Reorganize tests according to new endpoint structure
